Renamed files to follow new naming
