+++ src/pages/account/profile/index.tsx
@@ -24,11 +24,11 @@ import { sentryMiddleware } from "@src/middlewares/sentry-middleware";
 const ProfilePage: React.FC = () => {
   const { data, error } = useSWR<
     {
-      profileUserInfo: Profile;
-      profileAddresses: Address[];
+      userProfile: Profile;
+      userAddresses: Address[];
     },
     SWRError
-  >(`/api/profile/get-profile`);
+  >(`/api/profile/get-user-profile-and-addresses`);
 
   if (error) {
     throw error;
